YUK-15642 Go through current Data Streaming implementation and see if it makes sense  - Updated messages to differentiate between NM server and DB errors.

+++ yukon-client/common/src/com/cannontech/common/rfn/message/datastreaming/network/NetworkDataStreamingInfo.java
@@ -5,53 +5,56 @@ import java.io.Serializable;
 /**
  * Describes system configuration and network parameters for data streaming.
  * NM sends this information to Yukon in the following ways:
- * 1. Each time NM is started;
- * 2. When NM receives a RfnArchiveStartupNotification message, which indicates Yukon restarts;
+ * 1. Each time NM starts;
+ * 2. When NM receives a RfnArchiveStartupNotification message, which Yukon sends to NM each time Yukon starts;
  * 3. When NM receives a NetworkDataStreamingInfoRequest from Yukon.
  */
 public class NetworkDataStreamingInfo implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    private int unitReportingInterval; // to calculate the basic unit of DS capacity.  (e.x., 30 minutes)
+    private short unitReportingInterval; // to calculate the basic unit of DS capacity.  (e.x., 30 minutes)
     
-    private int minReportingInterval;  // Define the minimum reporting interval. (e.x.,  1 minute)
-    private int maxReportingInterval;  // Define the maximum reporting interval. (e.x., 30 minute)
+    private short minReportingInterval;  // Define the minimum reporting interval. (e.x.,  1 minute)
+    private short maxReportingInterval;  // Define the maximum reporting interval. (e.x., 30 minute)
     
-    private int maxNumberOfReportingIntervals; // Define the max number of reporting intervals you can configure for a device.
-                                               // For the initial phase, only one reporting interval is supported.
+    private int maxNumberOfReportingIntervals;
+        // Define the max number of reporting intervals you can configure for a device.
+        // For the initial phase, only one reporting interval is supported.
        
-    private int maxNumberOfMetricsPerReport; // Define the max number of DS metrics fitted in a report. For example, if the value
-                                             // is 8 and you already enabled one DS metric, the current firmware implementation
-                                             // supports you to add 7 more metrics without inputting additional loading.
+    private int maxNumberOfMetricsPerReport;
+        // Define the max number of DS metrics fitted in a report.
+        // Note load is calculated by reports.
+        // For example, if max metrics is set to 8 and you already enabled one DS metric,
+        // the current firmware implementation supports you to add 7 more metrics without inputting additional load.
     
-    private int maxNumberOfReportsPerDevice; // Define the max number of DS reports a device can send at a reporting time.
-                                             // So the max number of metrics you can configure for a device is the product of
-                                             // this value and the max number of DS metrics fitted in a report.
-                                             // For the initial phase, only one DS report per device is supported.
-                                             // So if maxNumberOfMetricsPerReport is 8, only 8 * 1 DS metrics
-                                             // can be configured for a device.
+    private int maxNumberOfReportsPerDevice;
+        // Define the max number of reports a device can send at a reporting time.
+        // Thus the max metrics you can enable for a device equals to
+        // the max metrics per report times max reports per device. For the initial phase,
+        // since only one report per device is supported, if the max metrics per report is 8,
+        // up to (8 * 1 = 8) metrics can be enabled for a device.
 
-    public int getUnitReportingInterval() {
+    public short getUnitReportingInterval() {
         return unitReportingInterval;
     }
 
-    public void setUnitReportingInterval(int unitReportingInterval) {
+    public void setUnitReportingInterval(short unitReportingInterval) {
         this.unitReportingInterval = unitReportingInterval;
     }
 
-    public int getMinReportingInterval() {
+    public short getMinReportingInterval() {
         return minReportingInterval;
     }
 
-    public void setMinReportingInterval(int minReportingInterval) {
+    public void setMinReportingInterval(short minReportingInterval) {
         this.minReportingInterval = minReportingInterval;
     }
 
-    public int getMaxReportingInterval() {
+    public short getMaxReportingInterval() {
         return maxReportingInterval;
     }
 
-    public void setMaxReportingInterval(int maxReportingInterval) {
+    public void setMaxReportingInterval(short maxReportingInterval) {
         this.maxReportingInterval = maxReportingInterval;
     }
